Add helpers if users selected wrong folder.
Many people are opening 1 folder too high. This PR causes that to be detected, and prompt that the right folder might be a subfolder.

If only one subfolder, only give the option to go to that.

If multiple, prompt allowing the user to pick.

Both options have a don't ask again.

Only goes 1 subfolder deep, not multiple.